Lumafuse Display Renderer overhaul
Using C++, multi-threading, compression, and native widget render targets, I've started laying the framework for a much more efficient and lower latency streaming solution.

The project is a little bit of a mess in this changelist and and needs to be cleaned up and refactored. Just submitting for peace of mind.